requirements.txt
requests
python-docx
google-generativeai
ttkbootstrap
customtkinter
